+++ src/koru/queue/__init__.py
@@ -3,6 +3,19 @@
 from koru.queue.human import default_human_prompt
 from koru.queue.koru_queue_argv import build_koru_queue_argv
 from koru.queue.loop import run_planfile_queue_loop
+from koru.proposal_envelope import NO_VALID_ARTIFACT
+from koru.queue.patch_mode import (
+    PATCH_DOES_NOT_APPLY,
+    POLICY_DENIED,
+    PROMOTION_CONFLICT,
+    UNSAFE_DIRTY_WORKSPACE,
+    VERIFY_BASELINE_FAILED,
+    VERIFY_FAILED_ISOLATED,
+    VERIFY_FAILED_ROLLED_BACK,
+    VERIFY_PROFILE_INVALID,
+    PatchOutcome,
+)
+from koru.queue.patch_transaction import PatchTransactionResult
 from koru.queue.runner import run_next_planfile_task
 from koru.queue.runners import (
     run_api_request,
@@ -20,6 +33,17 @@ from koru.queue.types import (
 )
 
 __all__ = [
+    "NO_VALID_ARTIFACT",
+    "PATCH_DOES_NOT_APPLY",
+    "POLICY_DENIED",
+    "PROMOTION_CONFLICT",
+    "UNSAFE_DIRTY_WORKSPACE",
+    "VERIFY_BASELINE_FAILED",
+    "VERIFY_FAILED_ISOLATED",
+    "VERIFY_FAILED_ROLLED_BACK",
+    "VERIFY_PROFILE_INVALID",
+    "PatchOutcome",
+    "PatchTransactionResult",
     "SHELL_RUN_NOTE_TAG",
     "build_koru_queue_argv",
     "format_shell_run_note",